Refaktorert ut inntektsperiode og vedtaksperiode fra vedtaksresultatswitch
