+++ src/frontend/komponenter/Behandling/VedtakOgBeregning/VedtakOgBeregning.tsx
@@ -5,10 +5,9 @@ import { AxiosRequestConfig } from 'axios';
 import DataViewer from '../../Felleskomponenter/DataViewer/DataViewer';
 import { useDataHenter } from '../../../hooks/felles/useDataHenter';
 import { AlertStripeFeil } from 'nav-frontend-alertstriper';
-import { EBehandlingResultat, IVedtak, IInntektsperiode } from '../../../typer/vedtak';
+import { EBehandlingResultat, IVedtak } from '../../../typer/vedtak';
 import VedtaksresultatSwitch from './VedtaksresultatSwitch';
 import VelgVedtaksresultat from './VelgVedtaksresultat';
-import { Checkbox } from 'nav-frontend-skjema';
 
 interface Props {
     behandlingId: string;
@@ -20,10 +19,6 @@ const StyledVedtaksperiode = styled.div`
 
 const StyledFeilmelding = styled(AlertStripeFeil)`
     margin-top: 2rem;
-`;
-
-const Inntekt = styled.div`
-    padding: 2rem;
 `;
 
 const VedtakOgBeregning: FC<Props> = ({ behandlingId }) => {
